Fixed bugs with headings
The \chapter heading was inside a group which prevented the @nobreak
flag from being set properly (it was being reset once the group exited).
This caused the spacing prior to the first section heading to be
incorrect.

The \part heading's end code was inserted at the wrong place and so
would cause problems if \part was following by anything other than
\chapter (causing a new page).